Cleaner code with addtocart mixin
