+++ src/themes/icmaa-imp/mixins/product/addtocartMixin.ts
@@ -1,0 +1,37 @@
+import { minValue } from 'vuelidate/lib/validators'
+import { notifications } from '@vue-storefront/core/modules/cart/helpers'
+
+export default {
+  methods: {
+    getQuantity () {
+      return this.$store
+        .dispatch('stock/check', {
+          product: this.product,
+          qty: this.product.qte
+        })
+        .then(res => {
+          this.quantity = res.qty
+        })
+    },
+    async addToCart (product) {
+      try {
+        const diffLog = await this.$store.dispatch('cart/addItem', { productToAdd: product })
+        diffLog.clientNotifications.forEach(notificationData => {
+          this.notifyUser(notificationData)
+        })
+      } catch (message) {
+        this.notifyUser(notifications.createNotification({ type: 'error', message }))
+      }
+    },
+    notifyUser (notificationData) {
+      this.$store.dispatch('notification/spawnNotification', notificationData, { root: true })
+    }
+  },
+  validations: {
+    product: {
+      qty: {
+        minValue: minValue(1)
+      }
+    }
+  }
+}
